% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cavi}
\alias{cavi}
\title{Variational inference for the sparse factor model}
\usage{
cavi(
  ymat,
  pivec,
  ptaushape,
  ptaurate,
  palphashape,
  palpharate,
  check = 100L,
  save = 0L,
  max_iter = 5000L,
  tol_elbo = 1e-10,
  tol_z = 0,
  seed = -1L
)
}
\arguments{
\item{ymat}{Data matrix, rows corresponding to features and columns corresponding to samples. May contain \code{NA}s.}

\item{pivec}{Vector of sparsity hyperparameters for each factor.}

\item{ptaushape}{Shape hyperparameter of the gamma prior for the feature-specific precision of the noise.}

\item{ptaurate}{Rate hyperparameter of the gamma prior for the feature-specific precision of the noise.}

\item{palphashape}{Shape hyperparameter of the gamma prior for the factor-specific precision of the loading factors.}

\item{palpharate}{Rate hyperparameter of the gamma prior for simulating the factor-specific precision of the loading factors.}

\item{check}{Convergence is checked once every \code{check} iterations. This is done because the ELBO is rather expensive to calculate.}

\item{save}{Variational parameters are stored (to be returned) once every \code{save} iterations. This is done to save memory. If \code{save} is \code{0}, only the final set of variational parameters is returned.}

\item{max_iter}{Maximum number of iterations to run.}

\item{tol_elbo}{Optimisation terminates when the difference in ELBO is less than \code{tol_elbo}.}

\item{tol_z}{Additional option for optimisation to terminate when the maximum difference in the variational parameter for \strong{Z} is less than \code{tol_z}.}

\item{seed}{Random seed. No seed is set when \code{seed} is \code{-1}.}
}
\value{
A list of variational parameters for each iteration saved.
\describe{
\item{lmean}{Mean of the normal variational approximation for the nonzero loading factors.}
\item{lsig}{Variance of the normal variational approximation for the nonzero loading factors.}
\item{fmean}{Mean of the normal variational approximation for the activation weights.}
\item{fsig}{Variance of the normal variational approximation for the activation weights.}
\item{zmean}{Mean of the Bernoulli variational approximation for the connectivity structure.}
\item{taushape}{Shape parameter of the gamma variational approximation for the feature-specific precision of the noise.}
\item{taurate}{Rate parameter of the gamma variational approximation for the feature-specific precision of the noise.}
\item{alphashape}{Shape parameter of the gamma variational approximation for the factor-specific precision of the loading factors.}
\item{alpharate}{Rate parameter of the gamma variational approximation for the factor-specific precision of the loading factors.}
\item{elbo}{Evidence lower bound.}
\item{iter}{Iteration number.}
\item{time}{Time when this iteration finished (in seconds).}
}
}
\description{
Runs coordinate ascent variational inference (CAVI) assuming a mean-field approximation.
}
\details{
The mean-field approximation is almost fully factorised, except that the dependency between corresponding entries of \strong{L} and \strong{Z} cannot be removed.
}
