% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cavi_fdep}
\alias{cavi_fdep}
\title{Variational inference for the sparse factor model with factor dependency for activation weights}
\usage{
cavi_fdep(
  ymat,
  pivec,
  ptaushape,
  ptaurate,
  palphashape,
  palpharate,
  check = 100L,
  save = 0L,
  max_iter = 5000L,
  tol_elbo = 1e-10,
  tol_z = 0,
  seed = -1L
)
}
\description{
Runs coordinate ascent variational inference (CAVI) assuming a mean-field approximation, where factor dependency is modelled for the activation weights.
}
\details{
Same functionality as \code{\link{cavi}}, except each column of the activation weights has a variational approximation, instead of each entry. These variational approximations are multivariate normal distributions, and share the same covariance matrix across columns. Dependency between factors is modelled, however, \code{ymat} must have no \code{NA}s.
}
