% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get.conf.mat}
\alias{get.conf.mat}
\title{Confusion matrix for connectivity structure predictions}
\usage{
get.conf.mat(true.z, zmeans)
}
\arguments{
\item{true.z}{Binary matrix of true connectivity structure used in data simulation.}

\item{zmeans}{Matrix of reals between 0 and 1 corresponding to the true connectivity strucutre. Predictions are taken by rounding each entry to the nearest integer.}
}
\value{
A 2-by-2 matrix of proportions of [[true positive, false negatives], [false positives, true negatives]].
}
\description{
Generates a confusion matrix consisting of proportions of true positives/negatives and false positives/negatives.
}
\details{
The order of factors must match, relabel first if needed.
}
