% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get.lferr}
\alias{get.lferr}
\title{Prediction errors of inferred matrices of the sparse factor model}
\usage{
get.lferr(true.l, true.f, lmeans, fmeans)
}
\arguments{
\item{true.l}{Matrix of loading factors used in data simulation.}

\item{true.f}{Matrix of factor activations used in data simulation.}

\item{lmeans}{Posterior mean of loading factors.}

\item{fmeans}{Posterior mean of factor activations.}
}
\value{
Vector of three relative root-mean-square errors for \strong{L},\strong{F},\strong{LF}.
}
\description{
Relative root-mean-square error of
}
\details{
The order of factors must match, relabel (including signflips) first if needed. Note that after relabelling, the likelihood is still invariant under rescaling of factors. For a fair comparison, rows of the posterior mean of \strong{F} are scaled to have the same norm as that of the true \strong{F}, and the columns of the posterior mean of \strong{L} are scaled correspondingly (an inverse scaling).
}
