% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get.zacc}
\alias{get.zacc}
\title{Accuracy of connectivity structure predictions}
\usage{
get.zacc(true.z, zmeans)
}
\arguments{
\item{true.z}{Binary matrix of true connectivity structure used in data simulation.}

\item{zmeans}{Matrix of reals between 0 and 1 corresponding to the true connectivity strucutre. Predictions are taken by rounding each entry to the nearest integer.}
}
\value{
The sum of true positives and true negatives (in proportions).
}
\description{
Proportion of correctly predicted entries of the connectivity structure
}
\details{
The order of factors must match, relabel first if needed.
}
