% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_relabelling}
\alias{get_relabelling}
\title{Relabelling factors for posterior summary to match a target}
\usage{
get_relabelling(fmeans, fsigs, fmat, print_mat = FALSE)
}
\arguments{
\item{fmeans}{Posterior mean of the activation matrix.}

\item{fsigs}{Posterior variance of the activation matrix.}

\item{fmat}{Target activation matrix.}

\item{print_mat}{Boolean for whether to print the cost matrix of the underlying linear assignment problem.}
}
\value{
A list of the permutation and signflips needed for the posterior summary to match the target. The permutation should be applied before the signflips.
\item{permutation}{Permutation (of 1 to K) to apply to the factors of the posterior summary.}
\item{sign}{Vector of 1s and -1s to multiply to the factors of the posterior summary after applying the permutation.}
}
\description{
Takes posterior mean and variance of the activation matrix and finds the relabelling needed for it to match a target (e.g. simualted dataset).
}
\details{
This is done by minimising a loss function via solving a linear assignment problem via the Jonker-Volgenant algorithm. The algorithm is implemented by Tomas Kazmar (https://github.com/gatagat/lap).
}
