% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs}
\alias{gibbs}
\title{entry point for MCMC, handles NAs}
\usage{
gibbs(
  n_samples,
  ymat,
  pivec,
  ptaushape,
  ptaurate,
  palphashape,
  palpharate,
  burn_in = 0L,
  thin = 1L,
  seed = -1L
)
}
\description{
entry point for MCMC, handles NAs
}
