% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myFunction}
\alias{myFunction}
\title{myFunction}
\usage{
myFunction(x)
}
\arguments{
\item{x}{a vector of strings}
}
\description{
Modify a string in Rcpp.
}
\examples{
myFunction(x=c('Hello', "C++", 'header', 'files'))

}
