% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relabel_params}
\alias{relabel_params}
\title{Relabel factors of a sample to match a target}
\usage{
relabel_params(params, fmeans, fsigs, print_mat = FALSE)
}
\arguments{
\item{params}{A list of model parameters. The naming follows the return value of \code{\link{sim.sfm}}.
\describe{
\item{lmat}{Matrix of loading factors.}
\item{fmat}{Matrix of activation weights.}
\item{zmat}{Binary matrix for the connectivity structure. Entries that are zero should enforce corresponding entries of \code{lmat} to be zero as well.}
\item{tauvec}{Vector of feature-specific precisions of the noise.}
\item{alphavec}{Vector of factor-specific precisions of the loading factors.}
}}

\item{print_mat}{Boolean for whether to print the cost matrix of the underlying linear assignment problem.}
}
\value{
A modified copy of \code{sample} with relabelled factors. Note that \code{ymat} is not returned even if provided in \code{params}.
}
\description{
Takes a set of model parameters and relabels the factors such that each entry of F resembles a normal distribution with target means and variances.
}
\details{
The negative log-likelihood is minimised by solving a linear assignment problem via the Jonker-Volgenant algorithm. The algorithm is implemented by Tomas Kazmar (https://github.com/gatagat/lap).
}
