% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relabel_samples}
\alias{relabel_samples}
\title{Relabel factors within MCMC samples}
\usage{
relabel_samples(
  samples,
  tol = 1e-08,
  print_action = FALSE,
  print_cost = FALSE,
  to_clone = TRUE
)
}
\arguments{
\item{samples}{A list of \eqn{T} MCMC samples as returned by \code{\link{gibbs}}.
\describe{
\item{lmat}{\eqn{T}-by-\eqn{G}-by-\eqn{K} array of the sampled loading factors.}
\item{fmat}{\eqn{T}-by-\eqn{K}-by-\eqn{N} array of the sampled activation weights.}
\item{zmat}{\eqn{T}-by-\eqn{G}-by-\eqn{K} array of the sampled connectivity structures.}
\item{tau}{\eqn{T}-by-\eqn{G} matrix of the sampled feature-specific precisions of the noise.}
\item{alpha}{\eqn{T}-by-\eqn{K} matrix of the sampled factor-specific precisions of the loading factors.}
\item{time}{Vector of sampling times of when samples were generated.}
}}

\item{tol}{Relabelling algorithm terminates when the change in likelihood is smaller than \code{tol}.}

\item{print_action}{Boolean for whether to print the final means and variances of the target normal distributions.}

\item{print_cost}{Boolean for whether to print the negative log-likelihood at each iteration.}

\item{to_clone}{Boolean for whether to copy \code{samples}.}
}
\value{
A modified copy of \code{samples} with relabelled factors.
}
\description{
Takes a list of samples and relabels the factors such that each entry of F resembles a normal distribution.
}
\details{
The negative log-likelihood is iteratively minimised by solving linear assignment problems via the Jonker-Volgenant algorithm. The algorithm is implemented by Tomas Kazmar (https://github.com/gatagat/lap).
}
