\name{sparsefactor-package}
\alias{sparsefactor-package}
\alias{sparsefactor}
\docType{package}
\title{\packageTitle{sparsefactor}}
\description{\packageDescription{sparsefactor}}
\section{Package Content}{\packageIndices{sparsefactor}}
\author{\packageAuthor{sparsefactor}}
\section{Maintainer}{\packageMaintainer{sparsefactor}}
\keyword{package}

