% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsefactor-package.R
\docType{package}
\name{sparsefactor}
\alias{sparsefactor}
\title{Bayesian Inference Techniques for Sparse Factor Models}
\description{
This packages provides two approaches to inference for sparse factor models: Gibbs sampling and variational inference. Additionally, a relabelling algorithm is implemented to address issues which arise from label switching and sign switching.
}
