% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cavi}
\alias{cavi}
\title{VI entry point, handles NAs
`check` is how often elbo is calculated and checked for convergence as full elbo calculation is expensive
`save` is how often the parameters are saved
algorithm terminates if max difference of z between two iterations is below `tol_z`
`tol_z` of 0 means the above is effectively not checked for'}
\usage{
cavi(
  ymat,
  pivec,
  ptaushape,
  ptaurate,
  palphashape,
  palpharate,
  check = 100L,
  save = 0L,
  max_iter = 5000L,
  tol_elbo = 1e-14,
  tol_z = 0,
  seed = -1L
)
}
\description{
VI entry point, handles NAs
`check` is how often elbo is calculated and checked for convergence as full elbo calculation is expensive
`save` is how often the parameters are saved
algorithm terminates if max difference of z between two iterations is below `tol_z`
`tol_z` of 0 means the above is effectively not checked for'
}
