% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{sim.sfm}
\alias{sim.sfm}
\title{Simulate for a sparse factor model}
\usage{
sim.sfm(
  lmat = NULL,
  tauvec = NULL,
  fmat = NULL,
  zmat = NULL,
  alphavec = NULL,
  G = 20,
  N = 10,
  K = 4,
  pivec = NULL,
  snr = NULL,
  taushape = 100,
  taurate = 1,
  alphashape = 1,
  alpharate = 1
)
}
\arguments{
\item{lmat}{Matrix of loading factors. If specified, \code{zmat} must also be specified. Otherwise, this will be simulated.}

\item{tauvec}{Vector of feature-specific precisions of the data. If \code{NULL}, this will be simulated.}

\item{fmat}{Matrix of factor activations. If \code{NULL}, this will be simulated.}

\item{zmat}{Binary matrix for the connectivity structure. Entries that are zero should enforce corresponding entries of \code{lmat} to be zero as well. If \code{NULL}, this will be simulated.}

\item{alphavec}{Vector of factor-specific precisions of the loading factors. If \code{NULL}, this will be simulated.}

\item{G}{Number of features.}

\item{N}{Number of samples.}

\item{K}{Number of factors.}

\item{pivec}{Factor-specific sparsity hyperparameters. If \code{NULL}, this will be set to 0.9 and 0.1 for the first half and second half of the vectors respectively.}

\item{snr}{Signal-to-noise ratio when simulating \eqn{\tau}. If \code{NULL}, \eqn{\tau} will instead be simulated from a gamma distribution.}

\item{taushape}{Shape parameter of the gamma distribution for simulating the feature-specific precision of the data.}

\item{taurate}{Rate parameter of the gamma distribution for simulating the feature-specific precision of the data.}

\item{alphashape}{Shape parameter of the gamma distribution for simulating the factor-specific precision of the loading factors.}

\item{alpharate}{Rate parameter of the gamma distribution for simulating the factor-specific precision of the loading factors.}
}
\value{
A list of all parameters (hyperparameters excluded), whether simulated or provided.
\itemize{
\item ymat - Data matrix with noise included.
\item lmat - Matrix of loading factors.
\item fmat - Matrix of factor activations.
\item zmat - Binary matrix for the connectivity structure. Entries that are zero should enforce corresponding entries of \code{lmat} to be zero as well.
\item tauvec - Vector of feature-specific precisions of the data.
\item alphavec -Vector of factor-specific precisions of the loading factors.
}
}
\description{
Simulates a dataset according to the sparse factor model
}
